removed duplicated code
