+++ frontend/src/components/EventViewSelection.tsx
@@ -1,4 +1,4 @@
-import React, {useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import {User} from "../model/User";
 import {EventModel} from "../model/EventModel";
 import Tabs from "react-bootstrap/Tabs";
@@ -17,6 +17,23 @@ type Props = {
 function EventViewSelection(props: Props) {
     const [activeEventFilter, setActiveEventFilter] = useState<string>("all");
 
+    useEffect(() => {
+        props.getEventsByStatus("APPROVED");
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, []);
+
+    function onAllEventsClickHandler() {
+        props.getEventsByStatus("APPROVED");
+        setActiveEventFilter("all");
+    }
+
+    function onMyEventsClickHandler() {
+        if (props.user?.preferredCategories) {
+            props.getEventsByCategory(props.user.preferredCategories);
+            setActiveEventFilter("my");
+        }
+    }
+
     return (
         <div>
             <Tabs
@@ -27,16 +44,17 @@ function EventViewSelection(props: Props) {
                 <Tab eventKey="Cards" title="Cards">
                     <Container>
                         <EventGallery user={props.user} events={props.events}
-                                      getEventsByStatus={props.getEventsByStatus}
-                                      getEventsByCategory={props.getEventsByCategory}
+                                      onAllEventsClickHandler={onAllEventsClickHandler}
+                                      onMyEventsClickHandler={onMyEventsClickHandler}
                                       activeEventFilter={activeEventFilter}
                                       setActiveEventFilter={setActiveEventFilter}/>
                     </Container>
                 </Tab>
 
                 <Tab eventKey="Carousel" title="Carousel">
-                    <EventCarousel user={props.user} events={props.events} getEventsByStatus={props.getEventsByStatus}
-                                   getEventsByCategory={props.getEventsByCategory} activeEventFilter={activeEventFilter}
+                    <EventCarousel user={props.user} events={props.events}
+                                   onAllEventsClickHandler={onAllEventsClickHandler}
+                                   onMyEventsClickHandler={onMyEventsClickHandler} activeEventFilter={activeEventFilter}
                                    setActiveEventFilter={setActiveEventFilter}/>
                 </Tab>
